fix some bug

+++ include/io_thread.h
@@ -6,8 +6,8 @@ typedef struct {
     int type;
     int fd;
     block_t *bt;
-    int offset;
-    int length;
+    off_t offset;
+    uint32_t length;
     char *buf;
     void *callback;
     void *data;
@@ -32,8 +32,8 @@ typedef struct {
 
 void io_thread_init(cycle_t *cycle);
 void main_distribute_block(cycle_t *cycle, int io_id, int block_id, int fd,
-        int offset, int length, block_t *bt, void *callback, void *data);
+        off_t offset, uint32_t length, block_t *bt, void *callback, void *data);
 void main_get_block(cycle_t *cycle, int io_id, int block_id, int fd,
-        int offset, int length, block_t *bt, void *callback, void *data);
+        off_t offset, uint32_t length, block_t *bt, void *callback, void *data);
 
 #endif
